corrección arranque

+++ symfony/build/entrypoint.sh
@@ -1,10 +1,7 @@
 #!/bin/bash
 
 #Borramos archivos para vaciar el directorio del proyecto y que comando composer no de fallo
-rm -f /$PROJNAME/.env
-rm -f /$PROJNAME/.gitignore
-
-[ -d /$PROJNAME/src ] || composer create-project symfony/skeleton:${SYMFONYV} $PROJNAME
+[ -d /$PROJNAME/src ] || rm -f /$PROJNAME/.env /$PROJNAME/.gitignore && composer create-project symfony/skeleton:${SYMFONYV} $PROJNAME
 
 cd /$PROJNAME
 symfony server:start
